refactor(SwaggerAuthToken): Swagger UI 에서 JWT 가 필요한 부분에 대한 설명 추가
