+++ src/main/java/roomescape/config/swagger/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package roomescape.config;
+package roomescape.config.swagger;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
