+++ src/main/java/roomescape/controller/ReservationController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.LoginMemberId;
+import roomescape.config.swagger.SwaggerAuthToken;
 import roomescape.service.reservation.ReservationService;
 import roomescape.service.reservation.dto.ReservationRequest;
 import roomescape.service.reservation.dto.ReservationResponse;
@@ -35,7 +36,7 @@ public class ReservationController {
     @PostMapping
     public ResponseEntity<ReservationResponse> createReservation(
             @RequestBody @Valid ReservationRequest reservationRequest,
-            @LoginMemberId long memberId
+            @LoginMemberId @SwaggerAuthToken long memberId
     ) {
         ReservationResponse reservationResponse = reservationService.create(reservationRequest, memberId);
         return ResponseEntity.created(URI.create("/reservations/" + reservationResponse.id()))
@@ -45,7 +46,7 @@ public class ReservationController {
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteReservation(
             @PathVariable("id") long reservationId,
-            @LoginMemberId long memberId
+            @LoginMemberId @SwaggerAuthToken long memberId
     ) {
         reservationService.deleteById(reservationId, memberId);
         return ResponseEntity.noContent().build();
